+++ xrlint/plugins/core/rules/content_desc.py
@@ -1,6 +1,6 @@
 import re
 
-from xrlint.node import DataArrayNode, DatasetNode
+from xrlint.node import VariableNode, DatasetNode
 from xrlint.plugins.core.plugin import plugin
 from xrlint.rule import RuleContext, RuleExit, RuleOp
 from xrlint.util.schema import schema
@@ -85,7 +85,7 @@ class ContentDesc(RuleOp):
             if attr_name not in dataset_attrs:
                 ctx.report(f"Missing attribute {attr_name!r}.")
 
-    def validate_data_array(self, ctx: RuleContext, node: DataArrayNode):
+    def validate_variable(self, ctx: RuleContext, node: VariableNode):
         if self.skip_vars:
             # Since dataset() has already been processed,
             # no need to check other nodes.
@@ -100,7 +100,7 @@ class ContentDesc(RuleOp):
                 # Ignored variable
                 return
 
-        var_attrs = node.data_array.attrs
+        var_attrs = node.array.attrs
         dataset_attrs = ctx.dataset.attrs
         for attr_name in self.common_attrs:
             if attr_name not in var_attrs and attr_name not in dataset_attrs: